Flatten and re-export.
