+++ README.md
@@ -1,2 +1,12 @@
 # Clipper2JS
-A port of Clipper2(C#) to js/ts
+A port of _[Clipper2](https://github.com/AngusJohnson/Clipper2)_(C#) to js/ts.
+
+## todo
+- [ ] Add build process.  
+- [ ] Add tests.  
+- [ ] Add comments.  
+- [ ] Add demo.  
+- [ ] Add benchmark.  
+
+## plan
+- TypedArray Path.
