+++ src/Engine/OutPt.ts
@@ -1,6 +1,6 @@
-import { Point64 } from "../Core/Point64";
 import { HorzSegment } from "./HorzSegment";
 import { OutRec } from "./OutRec";
+import { Point64 } from "../Core/Point64";
 
 export type OutPt = {
   pt: Point64;
