+++ src/Engine/OutRec.ts
@@ -1,8 +1,8 @@
-import { Path64 } from "../Core/Path64";
-import { Rect64 } from "../Core/Rect64";
 import { Active } from "./Active";
 import { OutPt } from "./OutPt";
 import { PolyPathBase } from "./PolyPathBase";
+import { Path64 } from "../Core/Path64";
+import { Rect64 } from "../Core/Rect64";
 
 export type OutRec = {
   idx: number;
