+++ src/Offset/ClipperGroup.ts
@@ -1,6 +1,6 @@
+import { EndType, JoinType } from "./OffsetEnums";
 import { Path64 } from "../Core/Path64";
 import { Paths64 } from "../Core/Paths64";
-import { EndType, JoinType } from "./OffsetEnums";
 
 export class ClipperGroup {
   inPaths: Paths64;
